requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1
pyyaml==6.0.2
python-multipart==0.0.20
